+++ packages/spacecat-shared-data-access/src/models/audit/audit.model.js
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-import { isArray, isObject } from '@adobe/spacecat-shared-utils';
+import { isArray, isObject, isBoolean } from '@adobe/spacecat-shared-utils';
 
 import { ValidationError } from '../../errors/index.js';
 import BaseModel from '../base/base.model.js';
@@ -47,6 +47,7 @@ class Audit extends BaseModel {
     FORMS_OPPORTUNITIES: 'forms-opportunities',
     SITE_DETECTION: 'site-detection',
     ALT_TEXT: 'alt-text',
+    ACCESSIBILITY: 'accessibility',
   };
 
   static AUDIT_TYPE_PROPERTIES = {
@@ -116,6 +117,7 @@ class Audit extends BaseModel {
        * @param {object[]} stepResult.urls - The list of URLs to scrape.
        * @param {string} stepResult.urls[].url - The URL to scrape.
        * @param {string} stepResult.siteId - The site ID. Will be used as the job ID.
+       * @param {string} stepResult.options - The options for the scraper.
        * @param {string} stepResult.processingType - The scraping processing type to trigger.
        * @param {object} auditContext - The audit context.
        * @param {object} context - The context object.
@@ -131,7 +133,8 @@ class Audit extends BaseModel {
         jobId: stepResult.siteId,
         processingType: stepResult.processingType || 'default',
         skipMessage: false,
-        allowCache: true,
+        allowCache: isBoolean(stepResult.allowCache) ? stepResult.allowCache : true,
+        options: stepResult.options || {},
         completionQueueUrl: stepResult.completionQueueUrl || context.env?.AUDIT_JOBS_QUEUE_URL,
         auditContext,
       }),