+++ packages/spacecat-shared-google-client/src/index.js
@@ -10,12 +10,12 @@
  * governing permissions and limitations under the License.
  */
 
-import AWSXray from 'aws-xray-sdk';
 import { google } from 'googleapis';
 import { OAuth2Client } from 'google-auth-library';
 import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
 import {
   composeAuditURL,
+  instrumentAWSClient,
   isArray,
   isInteger,
   isValidDate,
@@ -66,7 +66,7 @@ export default class GoogleClient {
     }
 
     const customerSecret = resolveCustomerSecretsName(baseURL, context);
-    const client = AWSXray.captureAWSv3Client(new SecretsManagerClient({}));
+    const client = instrumentAWSClient(new SecretsManagerClient({}));
 
     try {
       const command = new GetSecretValueCommand({ SecretId: customerSecret });