+++ packages/spacecat-shared-http-utils/src/index.d.ts
@@ -11,9 +11,11 @@
  */
 import { Response } from '@adobe/fetch';
 
-export declare function ok(body?: string): Response;
+export declare function ok(body?: string, headers?: object): Response;
 
-export declare function created(body: object): Response;
+export declare function created(body: object, headers?: object): Response;
+
+export declare function accepted(body: object, headers?: object): Response;
 
 export declare function noContent(headers?: object): Response;