+++ packages/spacecat-shared-http-utils/test/index.test.js
@@ -14,6 +14,7 @@ import { expect } from 'chai';
 import {
   ok, badRequest, notFound, internalServerError,
   noContent, found, created, createResponse, unauthorized, forbidden,
+  accepted,
 } from '../src/index.js';
 
 async function testMethod(response, expectedCode, expectedBody) {
@@ -53,10 +54,40 @@ describe('HTTP Response Functions', () => {
     await testMethod(response, 200, body);
   });
 
+  it('ok should return a 200 OK response with custom body and headers', async () => {
+    const body = { success: true };
+    const headers = { key: 'value' };
+    const response = await ok(body, headers);
+    await testMethod(response, 200, body);
+    expect(response.headers.get('key')).to.equal('value');
+  });
+
   it('created should return a 201 CREATED response with custom body', async () => {
     const body = { success: true };
     const response = await created(body);
     await testMethod(response, 201, body);
+  });
+
+  it('created should return a 200 OK response with custom body and headers', async () => {
+    const body = { success: true };
+    const headers = { key: 'value' };
+    const response = await created(body, headers);
+    await testMethod(response, 201, body);
+    expect(response.headers.get('key')).to.equal('value');
+  });
+
+  it('accepted should return a 202 ACCEPTED response with custom body', async () => {
+    const body = { status: 'ACCEPTED' };
+    const response = await accepted(body);
+    await testMethod(response, 202, body);
+  });
+
+  it('accepted should return a 200 OK response with custom body and headers', async () => {
+    const body = { status: 'ACCEPTED' };
+    const headers = { key: 'value' };
+    const response = await accepted(body, headers);
+    await testMethod(response, 202, body);
+    expect(response.headers.get('key')).to.equal('value');
   });
 
   it('noContent should return a 204 No Content response with default headers', async () => {